Add a check if the graphics card is nVidia.
If you don't an nVidia graphics card, no actions will be taken.

+++ install_nvidia_drivers.sh
@@ -1,10 +1,19 @@
 #!/usr/bin/env bash
 
-echo "Purge bumblebee and any existing nvidia drivers..."
-sudo apt-get purge libvdpau-va-gl1 bumblebee* nvidia*
+function install_nvidia() {
+    echo "1) Purge bumblebee and any existing nvidia drivers..."
+    sudo apt-get purge libvdpau-va-gl1 bumblebee* nvidia*
 
-echo "Install proprietary nvidia-331 drivers..."
-sudo apt-get install nvidia-331 nvidia-settings nvidia-prime prime-indicator
+    echo "2) Install proprietary nvidia-331 drivers..."
+    sudo apt-get install nvidia-331 nvidia-settings nvidia-prime prime-indicator
+}
 
-echo "Restart for the changes to take effect." 
-sudo shutdown -r 0
+if [[ `lspci | grep -i 'vga' | grep -i 'nvidia'` ]];
+then
+    echo "nVidia graphics card detected. Installing proprietary nVidia drivers..."
+    install_nvidia
+    echo "Restart for the changes to take effect." 
+    sudo shutdown -r 0
+else
+    echo "No nVidia graphics card detected."
+fi
